Limited number of Trail
This is my first attempt to make this game more interesting.
In this first try, I have limited the trail left behind. Interestingly
the trail left behind currenly is measured in number of turns (was
already there in the code, could not help it). So, for now the max trail
left is 10 turns.
Should make this configurable.