Changed the Logging mechanism
With the previous Logger, the className tag was needed to be set for
each class in it's constructor : inconvenient and too-much to remember.
So, made it now mandatory to pass the class when calling the Logger
methods. Also, added wrappers to the functions in case the programmer
gets as lazy as me and don't want to type the WHOLE name for the logging
functions.