commiting all the changes
The changes are all cosmetic. All I am trying to do is understand the
code so pepering it up with the log statements and indenting the code to
my taste. Git is still new to me, don't really know how to use it. So I
think all these changes are not relevent for the branch "trail", so want
to switch back to master. Hence, commiting all changes.