Refactored out the OpenGLRenderer class
Since the GLSurfaceView.Renderer is an interface and functionallity-vise
is very similar to a view (albit a bit more specific), I found it easier
to remove an extra layer of object-calling/message-passing
(whatever-it-was) to be... not necessery. Hence, merged the OpenGLView
and OpenGLRenderer classes and refactored (I have come to love that
term) the OpenGLRenderer class completely.